fix matrix3d int type error for windows (#81)

+++ llm/src/nn_modules/Fp32GPTBigCodeDecoder.cc
@@ -104,8 +104,8 @@ struct Fp32GPTBigCodeDecoder_output Fp32GPTBigCodeDecoder::forward(const struct 
     // Position embeddings
     // Matrix3D<float> pos_embeds = this->get_position_embed(sqlen, past_key_values_length);
 #ifdef _WIN32
-    std::vector<float> position_ids_buf_vec(sqlen);
-    float *position_ids_buf = &position_ids_buf_vec.front();
+    std::vector<int> position_ids_buf_vec(sqlen);
+    int *position_ids_buf = &position_ids_buf_vec.front();
     std::vector<float> pos_embeds_buf_vec(sqlen * this->embed_dim);
     float *pos_embeds_buf = &pos_embeds_buf_vec.front();
 #else
